+++ lib/response/pam_response.dart
@@ -27,8 +27,8 @@ class PamResponse {
     }
 
     var response = PamResponse();
-    final code = map['code'];
-    final errorMessage = map["message"];
+    final code = map['error_code'] ?? map['code'];
+    final errorMessage = map["error"] ?? map["message"];
 
     if (code != null) {
       response.error = PamErrorResponse(code: code, errorMessage: errorMessage);
@@ -36,7 +36,7 @@ class PamResponse {
 
     response.contactID = map["contact_id"];
     response.consentID = map["consent_id"];
-    response.database = map["database"];
+    response.database = map["_database"] ?? map["database"];
 
     return response;
   }